feat: 수강중인 과목 정보 저장 로직 구현
