+++ src/main/java/grit/guidance/domain/user/service/LoginService.java
@@ -8,6 +8,7 @@ import grit.guidance.domain.course.repository.TrackRepository;
 import grit.guidance.domain.user.dto.*;
 import grit.guidance.domain.user.entity.*;
 import grit.guidance.domain.user.repository.CompletedCourseRepository;
+import grit.guidance.domain.user.repository.EnrolledCourseRepository;
 import grit.guidance.domain.user.repository.UsersRepository;
 import grit.guidance.domain.user.repository.UserTrackRepository;
 import grit.guidance.global.jwt.JwtService;
@@ -32,6 +33,7 @@ public class LoginService {
     private final UsersRepository usersRepository;
     private final UserTrackRepository userTrackRepository;
     private final CompletedCourseRepository completedCourseRepository;
+    private final EnrolledCourseRepository enrolledCourseRepository;
     private final CourseRepository courseRepository;
     private final TrackRepository trackRepository;
     private final JwtService jwtService;
@@ -119,8 +121,12 @@ public class LoginService {
         
         // 3. 완료된 과목들 저장
         saveCompletedCourses(existingUser, hansungData.grades().semesters());
+
+        System.out.println(hansungData.enrolledCourseNames());
+        // 4. 수강 중인 과목들 저장
+        saveEnrolledCourses(existingUser, hansungData.enrolledCourseNames());
         
-        // 4. 사용자 lastCrawlTime과 updatedAt을 현재 시간으로 업데이트
+        // 5. 사용자 lastCrawlTime과 updatedAt을 현재 시간으로 업데이트
         existingUser.updateLastCrawlTime(); // 크롤링 시간 업데이트
         existingUser = usersRepository.save(existingUser); // @LastModifiedDate 트리거
         log.info("사용자 크롤링 시간 업데이트 완료: studentId={}, lastCrawlTime={}, updatedAt={}", 
@@ -343,5 +349,43 @@ public class LoginService {
             // 빈 문자열이거나 알 수 없는 경우 null 반환
             return null;
         }
+    }
+    
+    /**
+     * 수강 중인 과목들 저장
+     */
+    private void saveEnrolledCourses(Users user, List<String> enrolledCourseNames) {
+        if (enrolledCourseNames == null || enrolledCourseNames.isEmpty()) {
+            log.info("수강 중인 과목이 없습니다: userId={}", user.getId());
+            return;
+        }
+        
+        log.info("수강 중인 과목 저장 시작: userId={}, enrolledCourseNames={}", user.getId(), enrolledCourseNames);
+        
+        // 기존 수강 과목 데이터 삭제
+        enrolledCourseRepository.deleteByUser(user);
+        
+        int savedCount = 0;
+        int notFoundCount = 0;
+        
+        for (String courseName : enrolledCourseNames) {
+            var courseOpt = courseRepository.findByCourseName(courseName);
+            if (courseOpt.isPresent()) {
+                Course course = courseOpt.get();
+                EnrolledCourse enrolledCourse = EnrolledCourse.builder()
+                        .user(user)
+                        .course(course)
+                        .build();
+                enrolledCourseRepository.save(enrolledCourse);
+                savedCount++;
+                log.info("수강 과목 저장 완료: {} - {}", courseName, course.getCourseCode());
+            } else {
+                notFoundCount++;
+                log.warn("수강 과목을 DB에서 찾을 수 없음: {}", courseName);
+            }
+        }
+        
+        log.info("수강 과목 저장 완료 - 저장됨: {}, 찾을 수 없음: {}, 전체: {}", 
+                savedCount, notFoundCount, enrolledCourseNames.size());
     }
 }
